drivers/mxc/gpu-viv: Compiltion fix with gcc13
The nxp gpu-viv driver is falsely using gctBOOL in the header
as _QuerySignal's return type while the actual function returns gceSTATUS.
While both are of the same internal type, GCC-13 now checks the return type
to be of the same name and not only same representation and therefore fails
to compile the driver.
We fix it by correcting the return type in the header.

Upstream-status: No upstreamable, issue is with a downstream driver which
    is not merged upstream.

Signed-off-by: Mimoja <git@mimoja.de>