dts: Reenable wdog1 to allow for it to do the reset for us
With the fika v5 and v6 boards now being able to be used with
the display we can now start to adjust the dts to match the actual
pinouts. While we used the wdog1 physical pin on the symphony boards to
prototype the backlight control we are now no longer connecting the
backlight to those pins but the I2C4 pins instead.
By reenableing the watchdog1 we are also regaining the ability to
reboot as the variscite / nxp kernel is unable to reboot on its own.
Instead it fails to stop the watchdog which in return reboots the system
for us. Yiehks.

Upstream-status: Not upstreamable. Issue was introduced downstream and
    is only correlating to the 5.15 variscite / NXP kernel.

Signed-off-by: Mimoja <git@mimoja.de>